requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.7.7
Pygments==2.3.1
pytz==2019.3
PyYAML==5.1
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.2